Сделал загрузку образа и deploy Ollama идемпотентными

+++ scripts/03-download-cloud-image.sh
@@ -9,7 +9,10 @@ fi
 mkdir -p /var/lib/vz/template/qcow2
 cd /var/lib/vz/template/qcow2
 
-wget -O ubuntu-26.04.img \
-  https://cloud-images.ubuntu.com/releases/26.04/release/ubuntu-26.04-server-cloudimg-amd64.img
-
-echo "IMAGE DOWNLOADED"
+if [[ -f ubuntu-26.04.img ]]; then
+  echo "Образ ubuntu-26.04.img уже существует. Пропускаю повторную загрузку."
+else
+  wget -O ubuntu-26.04.img \
+    https://cloud-images.ubuntu.com/releases/26.04/release/ubuntu-26.04-server-cloudimg-amd64.img
+  echo "IMAGE DOWNLOADED"
+fi
